New Changes/Fix - Back-end revision
Releases for card, and non card related, creation is done, and the specific relationships were built as well to add values into balance or invoice.
Created a new table called "BalanceInvoice" to store the invoices that are linked to the specific balance, to use in queries.
Balance GET request now can list all balance as well as the "date_reference" passed in body request.
Balance/Card/Release creation was fixed.
Release deletion was fixed to delete values from balance and invoice.

https://trello.com/c/xUFkSqf0/10-revis%C3%A3o-do-c%C3%B3digo-back-end